+++ R/exportedMethods.R
@@ -555,7 +555,7 @@ unmakeMemoisable.default <- function(x) {
     }
   }
 
-  if (any(inherits(obj, "SpatVector"), inherits(obj, "SpatRaster"))) {
+  if (any(inherits(obj, c("SpatVector", "SpatRaster", "SpatExtent")))) {
     if (!requireNamespace("terra", quietly = TRUE))
       stop("Please install terra package")
     messageCache("...wrapping terra object for saving...", verboseLevel = 2, verbose = verbose)
@@ -563,31 +563,37 @@ unmakeMemoisable.default <- function(x) {
 
     # next is for terra objects --> terra::wrap is ridiculously slow for SpatVector objects; use
     #   custom version in reproducible where here
+    useWrap <- TRUE
     if (inherits(obj, "SpatRaster")) {
       if (all(nzchar(Filenames(obj)))) {
+        useWrap <- FALSE
         cls <- class(obj)
+        layerNams <- paste(names(obj), collapse = layerNamesDelimiter)
         obj2 <- asPath(Filenames(obj, allowMultiple = FALSE))
         obj <- asPath(Filenames(obj))
         attr(obj, "tags") <- c(attr(obj, "tags"),
                                paste0("origFilename:", basename2(obj)),
+                               paste0("origRelName:", makeRelative(obj, cachePath)),
                                paste0("origDirname:", dirname(obj)),
                                paste0("origGetWd:", getwd()),
+                               paste0("origCachePath:", cachePath),
                                paste0("fromDisk:", TRUE),
                                paste0("class:", cls),
                                paste0("fileFormat:", tools::file_ext(obj)),
                                paste0("saveRawFile:", TRUE),
                                paste0("loadFun:", "terra::rast"),
+                               paste0("layerNames:", layerNams),
                                paste0("whichFiles:", obj2)
                                )
-      } else {
-        obj <- terra::wrap(obj)
       }
-    } else {
-      if (missing(cachePath)) # not in Cache call
-        obj <- terra::wrap(obj)
-      else
-        obj <- wrapSpatVector(obj)
+    } else if (is(obj, "SpatVector") && !missing(cachePath)) {
+      useWrap <- FALSE
+      obj <- wrapSpatVector(obj)
     }
+
+    if (useWrap)
+      obj <- wrap(obj) # let method dispatch work
+
     attr(obj, ".Cache") <- attrs
 
     messageCache("\b Done!", verboseLevel = 2, verbose = verbose)
@@ -613,6 +619,16 @@ unmakeMemoisable.default <- function(x) {
   }
   if (any(inherits(obj, "data.table"))) {
     obj <- data.table::copy(obj)
+  }
+  if (is.character(obj)) {
+    tags <- attr(obj, "tags")
+    if (any(grepl("origFilename", tags))) {
+      tags1 <- parseTags(tags)
+      obj <- loadFile(tags1$tagValue[tags1$tagKey %in% "whichFiles"],
+                      fullCacheTableForObj = tags1, cachePath = cachePath)
+      names(obj) <- strsplit(tags1$tagValue[tags1$tagKey == "layerNames"],
+                             split = layerNamesDelimiter)[[1]]
+    }
   }
 
   obj
@@ -648,51 +664,63 @@ unmakeMemoisable.default <- function(x) {
 .dealWithClassOnRecovery.list <- function(obj, cachePath, cacheId,
                                          drv = getDrv(getOption("reproducible.drv", NULL)),
                                          conn = getOption("reproducible.conn", NULL)) {
-#   if (isOutputList || isOutputEnv) {
-    anyNames <- names(obj)
-    isSpatVector <- if (is.null(anyNames)) FALSE else all(names(obj) %in% spatVectorNamesForCache)
-    if (isTRUE(isSpatVector)) {
-      obj <- unwrapSpatVector(obj)
+  #   if (isOutputList || isOutputEnv) {
+  anyNames <- names(obj)
+  isSpatVector <- if (is.null(anyNames)) FALSE else all(names(obj) %in% spatVectorNamesForCache)
+  if (isTRUE(isSpatVector)) {
+    obj <- unwrapSpatVector(obj)
+  } else {
+    if (!"cacheRaster" %in% names(obj)) { # recursive up until a list has cacheRaster name
+
+      outList <- obj
+      outList <- lapply(outList, function(out) .dealWithClassOnRecovery(out, cachePath, cacheId,
+                                                                        drv, conn))
+
+      obj <- outList
     } else {
-      if (!"cacheRaster" %in% names(obj)) { # recursive up until a list has cacheRaster name
-
-        outList <- obj
-        outList <- lapply(outList, function(out) .dealWithClassOnRecovery(out, cachePath, cacheId,
-                                                                         drv, conn))
-
-          obj <- outList
+      origFilenames <- if (is(obj, "Raster")) {
+        Filenames(obj) # This is legacy piece which allows backwards compatible
       } else {
-        origFilenames <- if (is(obj, "Raster")) {
-          Filenames(obj) # This is legacy piece which allows backwards compatible
+        obj$origRaster
+      }
+
+      filesExist <- file.exists(origFilenames)
+      cacheFilenames <- Filenames(obj)
+      filesExistInCache <- file.exists(cacheFilenames)
+      if (any(!filesExistInCache)) {
+        fileTails <- gsub("^.+(rasters.+)$", "\\1", cacheFilenames)
+        correctFilenames <- file.path(cachePath, fileTails)
+        filesExistInCache <- file.exists(correctFilenames)
+        if (all(filesExistInCache)) {
+          cacheFilenames <- correctFilenames
         } else {
-          obj$origRaster
+          stop("File-backed raster files in the cache are corrupt for cacheId: ", cacheId)
         }
-
-        filesExist <- file.exists(origFilenames)
-        cacheFilenames <- Filenames(obj)
-        filesExistInCache <- file.exists(cacheFilenames)
-        if (any(!filesExistInCache)) {
-          fileTails <- gsub("^.+(rasters.+)$", "\\1", cacheFilenames)
-          correctFilenames <- file.path(cachePath, fileTails)
-          filesExistInCache <- file.exists(correctFilenames)
-          if (all(filesExistInCache)) {
-            cacheFilenames <- correctFilenames
-          } else {
-            stop("File-backed raster files in the cache are corrupt for cacheId: ", cacheId)
-          }
-
-        }
-        out <- hardLinkOrCopy(cacheFilenames[filesExistInCache],
-                              origFilenames[filesExistInCache], overwrite = TRUE)
-
-        newOutput <- updateFilenameSlots(obj$cacheRaster,
-                                         Filenames(obj, allowMultiple = FALSE),
-                                         newFilenames = grep("\\.gri$", origFilenames, value = TRUE, invert = TRUE))
-        obj <- newOutput
-        obj <- .setSubAttrInList(obj, ".Cache", "newCache", FALSE)
-        obj
 
       }
+      out <- hardLinkOrCopy(cacheFilenames[filesExistInCache],
+                            origFilenames[filesExistInCache], overwrite = TRUE)
+
+      newOutput <- updateFilenameSlots(obj$cacheRaster,
+                                       Filenames(obj, allowMultiple = FALSE),
+                                       newFilenames = grep("\\.gri$", origFilenames, value = TRUE, invert = TRUE))
+      obj <- newOutput
+      obj <- .setSubAttrInList(obj, ".Cache", "newCache", FALSE)
+      obj
+
     }
+  }
   # }
+}
+
+
+parseTags <- function(tags) {
+  out <- strsplit(tags,':')
+  tags2 <- lapply(out, function(x) x[[1]])
+  tags1 <- as.data.table(do.call(rbind, tags2))
+  vals <- vapply(seq(NROW(tags1)), function(i)
+    gsub(paste0(tags1[[1]][i], ":"), "", tags[i]), FUN.VALUE = character(1))
+  set(tags1, NULL, "tagValue", vals)
+  setnames(tags1, "V1", "tagKey")
+  tags1
 }