+++ tests/testthat/test-checkPath.R
@@ -41,7 +41,7 @@ test_that("checkPath: normPath consistency", {
 })
 
 test_that("checkPath: checkPath consistency", {
-  testInitOut <- testInit()
+  testInit()
   withr::local_dir(tmpdir)
   # on.exit({
   #   testOnExit(testInitOut)